Just use unsigned int to match the coalescer->forwarder protocol

+++ ack-coalescing/ack_coalescer/coalescer.h
@@ -8,7 +8,7 @@
 #include <unistd.h>
 
 struct packet {
-        ssize_t len;
+        unsigned int len;
 	struct iphdr *ip;
 	struct tcphdr *tcp;
 	struct packet *next;
